+++ .gitignore
@@ -1,8 +1,21 @@
 driver/
 chromedriver/
+chromedriver/.DS_Store
 selenium-server-standalone-3.141.59.jar
 var/
+var/.DS_Store
 cred/
+cred/.DS_Store
 selenium/
+selenium/.DS_Store
 .DS_Store
 .vscode/
+__pycache__/
+utility/__pycache__/
+utility/.DS_Store
+mac/__pycache__/
+mac/.DS_Store
+windows/__pycache__/
+windows/.DS_Store
+linux/__pycache__/
+linux/.DS_Store
