+++ cleanup.py
@@ -1,0 +1,4 @@
+import os
+from path import  var_dir, cred_dir
+os.system("rm -rf " + var_dir)
+os.system("rm -rf " + cred_dir)
