+++ path.py
@@ -1,0 +1,4 @@
+import os
+path = os.getcwd()
+var_dir = os.path.join(path, "var")
+cred_dir = os.path.join(path, "cred")
